\name{mdata}
\alias{mergedata}
\alias{chen}
\alias{chen.poe}
\alias{chen.spl}
\alias{garber}
\alias{garber.poe}
\alias{garber.spl}
\alias{lapointe}
\alias{lapointe.poe}
\alias{lapointe.spl}
\title{metaArray sample dataset}
\description{ Three datasets from liver, lung, and prostate cancer microarrays. 
Please refer to the bibliography in the vignette. Chen (30 primary, 9 metastatic), 
Garber (30 primary, 6 metastatic), Lapointe (30 primary, 9 metastatic)}
\usage{data(mdata)}

\keyword{methods}
